DX-2184: Using 2.0.0 grindKey as the default version
